[ICU-900] Properly fix a compiler warning for Solaris X-Commit-URL: https://ssl.icu-project.org/trac/changeset/8403 Former-commit-id: 8fb737412314d3fd4f0c503756aa387417e74a6f

+++ icu4c/source/test/cintltst/ucmptst.c
@@ -94,7 +94,7 @@ static void TestUCMP8API(){
         log_err("Error in ucmp8_getkBlockCount()\n");
     } 
     values=(int8_t*)ucmp8_getArray(&ucmp8Array1);
-    if(values[0] !=TEST_DEFAULT_VALUE){
+    if((uint8_t)values[0] != (uint8_t)TEST_DEFAULT_VALUE){
         log_err("Error: getArray() or init failed\n");
     }
